refactor: Grouped model pages in folder, updated imports
