--- src/pages/Compute/Compute.tsx
@@ -1,7 +1,0 @@
-export const Compute = () => {
-  return (
-    <div>
-      <p>Compute models here!</p>
-    </div>
-  )
-}
