+++ src/pages/ModelPages/Compute/Compute.tsx
@@ -1,0 +1,7 @@
+export const Compute = () => {
+  return (
+    <>
+      <p>Options to compute models will soon be possible here!</p>
+    </>
+  )
+}
