+++ src/pages/ModelPages/Model/Model.styled.tsx
