+++ src/pages/ModelPages/Results/Results.tsx
@@ -1,0 +1,7 @@
+export const Results = () => {
+  return (
+    <>
+      <p>Results for models will soon be available here!</p>
+    </>
+  )
+}
