--- src/pages/Results/Results.tsx
@@ -1,7 +1,0 @@
-export const Results = () => {
-  return (
-    <div>
-      <p>Resukts for models here!</p>
-    </div>
-  )
-}
